program fails

+++ zhuravlev.ivan/I2/cnt.max.elem.cpp
@@ -2,8 +2,8 @@
 
 size_t CounterMaxElement(const int *arr, const int size)
 {
-  size_t cnt_max = 0;
-  size_t max = 0;
+  int cnt_max = 0;
+  int max = 0;
   for (int i = 0; i < size; ++i)
   {
     if (arr[i] > max)
