+++ zhuravlev.ivan/I2/left.shift.mover.cpp
@@ -3,7 +3,7 @@
 
 int* LeftShiftMover(int *arr, const size_t size,const size_t wanted_shift)
 {
-  int current_shifts = 0;
+  size_t current_shifts = 0;
   while (current_shifts++ < wanted_shift)
   {
     int temp = arr[0];
